--- a/scipy/optimize/meson.build
+++ b/scipy/optimize/meson.build
@@ -61,16 +61,16 @@ _minpack = py3.extension_module('_minpack',
   subdir: 'scipy/optimize'
 )
 
-rectangular_lasp = static_library('rectangular_lasp',
+rectangular_lsap = static_library('rectangular_lsap',
   [
     'rectangular_lsap/rectangular_lsap.h',
     'rectangular_lsap/rectangular_lsap.cpp'
   ],
 )
 
-py3.extension_module('_lsap_module',
-  ['_lsap_module.c'],
-  link_with: rectangular_lasp,
+py3.extension_module('_lsap',
+  ['_lsap.c'],
+  link_with: rectangular_lsap,
   c_args: numpy_nodepr_api,
   dependencies: [py3_dep],
   include_directories: include_dirs,
@@ -293,7 +293,6 @@ py3.install_sources([
   '_linprog_rs.py',
   '_linprog_simplex.py',
   '_linprog_util.py',
-  '_lsap.py',
   '_milp.py',
   '_minimize.py',
   '_minpack_py.py',