--- a/scipy/optimize/rectangular_lsap/rectangular_lsap.cpp
+++ b/scipy/optimize/rectangular_lsap/rectangular_lsap.cpp
@@ -57,18 +57,18 @@ template <typename T> std::vector<intptr_t> argsort_iter(const std::vector<T> &v
 }
 
 static intptr_t
-augmenting_path(intptr_t nc, std::vector<double>& cost, std::vector<double>& u,
+augmenting_path(intptr_t nc, double *cost, std::vector<double>& u,
                 std::vector<double>& v, std::vector<intptr_t>& path,
                 std::vector<intptr_t>& row4col,
                 std::vector<double>& shortestPathCosts, intptr_t i,
-                std::vector<bool>& SR, std::vector<bool>& SC, double* p_minVal)
+                std::vector<bool>& SR, std::vector<bool>& SC,
+                std::vector<intptr_t>& remaining, double* p_minVal)
 {
     double minVal = 0;
 
     // Crouse's pseudocode uses set complements to keep track of remaining
     // nodes.  Here we use a vector, as it is more efficient in C++.
     intptr_t num_remaining = nc;
-    std::vector<intptr_t> remaining(nc);
     for (intptr_t it = 0; it < nc; it++) {
         // Filling this up in reverse order ensures that the solution of a
         // constant cost matrix is the identity matrix (c.f. #11602).
@@ -107,11 +107,11 @@ augmenting_path(intptr_t nc, std::vector<double>& cost, std::vector<double>& u,
         }
 
         minVal = lowest;
-        intptr_t j = remaining[index];
         if (minVal == INFINITY) { // infeasible cost matrix
             return -1;
         }
 
+        intptr_t j = remaining[index];
         if (row4col[j] == -1) {
             sink = j;
         } else {
@@ -120,7 +120,6 @@ augmenting_path(intptr_t nc, std::vector<double>& cost, std::vector<double>& u,
 
         SC[j] = true;
         remaining[index] = remaining[--num_remaining];
-        remaining.resize(num_remaining);
     }
 
     *p_minVal = minVal;
@@ -128,7 +127,7 @@ augmenting_path(intptr_t nc, std::vector<double>& cost, std::vector<double>& u,
 }
 
 static int
-solve(intptr_t nr, intptr_t nc, double* input_cost, bool maximize,
+solve(intptr_t nr, intptr_t nc, double* cost, bool maximize,
       int64_t* a, int64_t* b)
 {
     // handle trivial inputs
@@ -136,35 +135,40 @@ solve(intptr_t nr, intptr_t nc, double* input_cost, bool maximize,
         return 0;
     }
 
-    std::vector<double> cost(input_cost, input_cost + nr * nc);
-
     // tall rectangular cost matrix must be transposed
     bool transpose = nc < nr;
-    if (transpose) {
-        for (intptr_t i = 0; i < nr; i++) {
-            for (intptr_t j = 0; j < nc; j++) {
-                cost[j * nr + i] = input_cost[i * nc + j];
+
+    // make a copy of the cost matrix if we need to modify it
+    std::vector<double> temp;
+    if (transpose || maximize) {
+        temp.resize(nr * nc);
+
+        if (transpose) {
+            for (intptr_t i = 0; i < nr; i++) {
+                for (intptr_t j = 0; j < nc; j++) {
+                    temp[j * nr + i] = cost[i * nc + j];
+                }
+            }
+
+            std::swap(nr, nc);
+        }
+        else {
+            std::copy(cost, cost + nr * nc, temp.begin());
+        }
+
+        // negate cost matrix for maximization
+        if (maximize) {
+            for (intptr_t i = 0; i < nr * nc; i++) {
+                temp[i] = -temp[i];
             }
         }
 
-        std::swap(nr, nc);
+        cost = temp.data();
     }
 
-    // negate cost matrix for maximization
-    if (maximize) {
-        for (intptr_t i = 0; i < nr * nc; i++) {
-            cost[i] = -cost[i];
-        }
-    }
-
-    // build a non-negative cost matrix
-    double minval = *std::min_element(cost.begin(), cost.end());
+    // test for NaN and -inf entries
     for (intptr_t i = 0; i < nr * nc; i++) {
-        auto v = cost[i] - minval;
-        cost[i] = v;
-
-        // test for NaN and -inf entries
-        if (v != v || v == -INFINITY) {
+        if (cost[i] != cost[i] || cost[i] == -INFINITY) {
             return RECTANGULAR_LSAP_INVALID;
         }
     }
@@ -178,13 +182,15 @@ solve(intptr_t nr, intptr_t nc, double* input_cost, bool maximize,
     std::vector<intptr_t> row4col(nc, -1);
     std::vector<bool> SR(nr);
     std::vector<bool> SC(nc);
+    std::vector<intptr_t> remaining(nc);
 
     // iteratively build the solution
     for (intptr_t curRow = 0; curRow < nr; curRow++) {
 
         double minVal;
         intptr_t sink = augmenting_path(nc, cost, u, v, path, row4col,
-                                        shortestPathCosts, curRow, SR, SC, &minVal);
+                                        shortestPathCosts, curRow, SR, SC,
+                                        remaining, &minVal);
         if (sink < 0) {
             return RECTANGULAR_LSAP_INFEASIBLE;
         }