… scipy/optimize/_lsap.py